making general organization and working on controller package

+++ service/oddService.go
@@ -1,41 +1,64 @@
 package service
 
 import (
+	"errors"
 	"log"
 	"sync"
 	"time"
 
+	"github.com/go-playground/validator/v10"
 	"gitthub.com/dionisiopro/dobet/api"
 	"gitthub.com/dionisiopro/dobet/models"
 	"gitthub.com/dionisiopro/dobet/repository"
 )
 
-var OddService = &oddService{}
+type OddService interface {
+	UpSertOdd(odd models.Odds) error
+	GetOddByMatchId(matchid string) (models.Odds, error )
+	UpdateOdds(leagueId string) error
+	DeleteOdd(odd_id string) error
+}
 
 type oddService struct {
 	repository repository.OddRepository
 	footballpi api.FootBallApi
+	leagueservice LeagueService
 }
 
-func SetUpOddServivce(repository repository.OddRepository,footballpi api.FootBallApi ){
-	OddService.footballpi = footballpi
-	OddService.repository = repository
-	lunchUpdateOddsLoop()
+func SetUpOddServivce(repository repository.OddRepository, footballpi api.FootBallApi, 	leagueservice LeagueService) OddService {
+	return &oddService{
+		repository: repository,
+		footballpi: footballpi,
+	}
 }
 
 func (service *oddService) UpSertOdd(odd models.Odds) error {
+	validate := validator.New()
+	err := validate.Struct(odd)
+	if err != nil{
+		return err
+	}
 	return service.repository.UpSertOdd(odd)
 
 }
-func (service *oddService) GetOddByMatchId(matchid string) (models.Odds){
+func (service *oddService) GetOddByMatchId(matchid string)( models.Odds, error ){
+	if matchid == ""{
+		return models.Odds{}, errors.New("Match Id Invalid")
+	}
 	return service.repository.GetOddByMatchId(matchid)
 }
 func (service *oddService) DeleteOdd(odd_id string) error {
+	if odd_id == ""{
+		return errors.New("Invalid odd id")
+	}
 	return service.repository.DeleteOdd(odd_id)
 }
 
 func (service *oddService) UpdateOdds(leagueId string) error {
-	oddDto, err := OddService.footballpi.GetOddsByLeagueId(leagueId)
+	if leagueId == ""{
+		return errors.New("Invalid league id")
+	}
+	oddDto, err := service.footballpi.GetOddsByLeagueId(leagueId)
 	if err != nil {
 		return err
 	}
@@ -47,25 +70,25 @@ func (service *oddService) UpdateOdds(leagueId string) error {
 	for _, odd := range odds {
 		func(odd models.Odds, wg *sync.WaitGroup) {
 			defer wg.Done()
-			OddService.repository.UpSertOdd(odd)
+			service.repository.UpSertOdd(odd)
 		}(odd, wg)
 	}
 	wg.Wait()
 	return nil
 }
 
-func lunchUpdateOddsLoop() {
+func (service *oddService) LunchUpdateOddsLoop() {
 	tiker := time.NewTicker(time.Hour * 24)
 	leagues := []models.League{}
 
 	for tker := range tiker.C {
 		if len(LocalLeagues) == 0 {
-			localLeagues, err := LeagueService.Leagues(0, 0)
+			localLeagues, err := service.leagueservice.Leagues(0, 0)
 			leagues = localLeagues
 			if err != nil {
 				return
 			}
-			
+
 		} else {
 			for _, league := range LocalLeagues {
 				leagues = append(leagues, league)
@@ -75,7 +98,7 @@ func lunchUpdateOddsLoop() {
 		for _, league := range leagues {
 			log := log.Default()
 			log.Print(tker)
-			go OddService.UpdateOdds(league.League_id)
+			go service.UpdateOdds(league.League_id)
 		}
 
 	}
